requirements.txt
Flask>=0.8
kafka-python==1.3.2
mock==2.0.0
nose==1.3.7
nose-parameterized==0.5.0
PyDispatcher==2.0.5
Pypubsub==3.3.0
PyYAML==3.12
twilio>=3.3.6
